added new excpetion to errors module. added some TODOs
