+++ src/compas_timber/errors/__init__.py
@@ -1,6 +1,8 @@
 class FeatureApplicationError(Exception):
     """Raised when a feature cannot be applied to an element geometry.
 
+    # TODO: perhaps should be renamed to ProcessingVisualizationError or something similar.
+
     Attributes
     ----------
     feature_geometry : :class:`~compas.geometry.Geometry`
@@ -67,8 +69,43 @@ class FastenerApplicationError(Exception):
         self.message = message
 
 
+class BTLxProcessingError(Exception):
+    """Exception raised when an error occurs while writing a Processing to BTLx file.
+
+    TODO: some work here to figure out the different types of feature/processing related errors.
+    TODO: this one is somewhat similar to FeatureApplicationError, but only relevant when processing is created from its proxy.
+    TOOD: also BTLxProcessingError is never throws but rather collected to form some sort of a report for the user.
+
+    Parameters
+    ----------
+    message : str
+        The error message.
+    part : :class:`BTLxPart`
+        The part that caused the error.
+    failed_processing : :class:`BTLxProcessing`
+        The processing that caused the error.
+
+    Attributes
+    ----------
+    message : str
+        The error message.
+    part : :class:`BTLxPart`
+        The part that caused the error.
+    failed_processing : :class:`BTLxProcessing`
+        The processing that caused the error.
+
+    """
+
+    def __init__(self, message, part, failed_processing):
+        super(BTLxProcessingError, self).__init__(message)
+        self.message = message
+        self.part = part
+        self.failed_processing = failed_processing
+
+
 __all__ = [
     "FeatureApplicationError",
     "BeamJoiningError",
     "FeatureApplicationError",
+    "BTLxProcessingError",
 ]
